pyproject.toml
[project]
name = "autourgos-google-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]
